Apricot's new Http service is separated from the already existing service.
