+++ apricot/apricot.go
@@ -46,6 +46,7 @@ func Run() (err error) {
 	log.Infof("AliECS Configuration Service (apricot v%s build %s) starting up", product.VERSION, product.BUILD)
 
 	s := remote.NewServer(Instance())
+	local.NewHttpService(Instance())
 	signals(s) // handle UNIX signals
 	var lis net.Listener
 	lis, err = net.Listen("tcp", fmt.Sprintf(":%d", viper.GetInt("listenPort")))
